Выносит утилиты и константы в отдельные файлы
