+++ src/utils.js
@@ -1,0 +1,11 @@
+const createTemplateElement = (template) => {
+  const templateElement = document.createElement(`template`);
+  templateElement.innerHTML = template;
+  return templateElement;
+};
+
+const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export {createTemplateElement, render};
